requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-bcrypt
python-dotenv
